handle run.py within a container

+++ run.py
@@ -3,8 +3,9 @@
 import argparse
 from distutils.spawn import find_executable
 from glob import glob
+import json
 from kvirt.config import Kconfig
-from kvirt.common import fetch, pprint, get_parameters
+from kvirt.common import pprint, get_parameters
 import os
 import re
 from shutil import copy2, rmtree
@@ -20,11 +21,31 @@ virtplatforms = ['kvm', 'kubevirt', 'ovirt', 'openstack', 'vsphere']
 cloudplatforms = ['aws', 'gcp']
 
 
-def get_path(x):
+def code_path(x):
+    if x is None:
+        return None
+    result = '/%s' % x if os.path.exists('/i_am_a_container') else './%s' % x
+    return result
+
+
+def pwd_path(x):
     if x is None:
         return None
     result = '/workdir/%s' % x if os.path.exists('/i_am_a_container') else x
     return result
+
+
+def ssh_path(x):
+    if x is None:
+        return None
+    elif not os.path.exists('/i_am_a_container'):
+        return ''
+    elif os.path.exists('/root/.ssh'):
+        return '/root/.ssh'
+    elif os.path.exists('/root/.kcli'):
+        return '/root/.kcli'
+    else:
+        return ''
 
 
 def real_path(x):
@@ -43,6 +64,44 @@ def get_values(data, element, field):
         new = data['%s_%s' % (element, field)]
         results.extend(new)
     return results
+
+
+def get_installer(nightly=False):
+    repo = 'ocp-dev-preview' if nightly else 'ocp'
+    INSTALLSYSTEM = 'mac' if os.path.exists('/Users') else 'linux'
+    msg = 'Downloading latest openshift-install from'
+    'https://mirror.openshift.com/pub/openshift-v4/clients/%s in current directory' % repo
+    pprint(msg, color='blue')
+    r = urlopen("https://mirror.openshift.com/pub/openshift-v4/clients/%s/latest/release.txt" % repo).readlines()
+    version = None
+    for line in r:
+        if 'Name' in str(line):
+            version = str(line).split(':')[1].strip().replace('\\n', '').replace("'", "")
+            break
+    if version is None:
+        pprint("Coudldn't find version", color='red')
+        os._exit(1)
+    cmd = "curl -s https://mirror.openshift.com/pub/openshift-v4/clients/%s/latest/" % repo
+    cmd += "openshift-install-%s-%s.tar.gz " % (INSTALLSYSTEM, version)
+    cmd += "| tar zxvf - %s/openshift-install" % pwd_path('.')
+    call(cmd, shell=True)
+
+
+def get_upstream_installer():
+    INSTALLSYSTEM = 'mac' if os.path.exists('/Users') else 'linux'
+    msg = 'Downloading latest okd openshift-install from github in current directory'
+    pprint(msg, color='blue')
+    r = urlopen("https://api.github.com/repos/openshift/okd/releases")
+    data = json.loads(r.read())
+    version = sorted([x['tag_name'] for x in data])[-1]
+    cmd = "curl -Ls https://github.com/openshift/okd/releases/download/"
+    cmd += "%s/openshift-install-%s-%s.tar.gz" % (version, INSTALLSYSTEM, version)
+    cmd += "| tar zxvf - %s/openshift-install" % pwd_path('.')
+    call(cmd, shell=True)
+
+
+def get_stable_installer():
+    print("prout")
 
 
 def gather_dhcp(data, platform):
@@ -90,11 +149,14 @@ def gather_dhcp(data, platform):
         return {'node_names': node_names, 'node_macs': node_macs, 'node_ips': node_ips, 'nodes': nodes}
 
 
-def template():
-    pprint("Generating a sample parameters.yml file", color='green')
-    params = get_parameters(get_path('ocp.yml'))
+def template(paramfile):
+    if os.path.exists(pwd_path(paramfile)):
+        pprint("Specified parameter file %s already there.Leaving..." % real_path(paramfile), color='red')
+        sys.exit(1)
+    pprint("Generating parameter file %s" % real_path(paramfile), color='green')
+    params = get_parameters(code_path('ocp.yml'))
     parameters = '\n'.join([parameter.lstrip() for parameter in params.split('\n')[1:]])
-    path = get_path('parameters.yml')
+    path = pwd_path(paramfile)
     with open(path, 'w') as f:
         f.write(parameters)
 
@@ -125,9 +187,9 @@ def scale(paramfile, workers):
     paramdata['scale'] = True
     paramdata['workers'] = workers
     if platform in virtplatforms:
-        config.plan(cluster, inputfile=get_path('ocp.yml'), overrides=paramdata)
+        config.plan(cluster, inputfile=code_path('ocp.yml'), overrides=paramdata)
     elif platform in cloudplatforms:
-        config.plan(cluster, inputfile=get_path('ocp_cloud.yml'), overrides=paramdata)
+        config.plan(cluster, inputfile=code_path('ocp_cloud.yml'), overrides=paramdata)
 
 
 def clean(paramfile):
@@ -144,7 +206,7 @@ def clean(paramfile):
         paramdata = {}
     cluster = paramdata.get('cluster', 'testk')
     config.plan(cluster, delete=True)
-    clusterdir = get_path("clusters/%s" % cluster)
+    clusterdir = pwd_path("clusters/%s" % cluster)
     if os.path.exists(clusterdir):
         pprint("Deleting %s" % real_path(clusterdir), color='green')
         rmtree(clusterdir)
@@ -152,7 +214,6 @@ def clean(paramfile):
 
 def deploy(paramfile):
     SYSTEM = 'macosx' if os.path.exists('/Users') else 'linux'
-    # INSTALLSYSTEM = 'macosx' if os.path.exists('/Users') else 'linux'
     config = Kconfig()
     k = config.k
     client = config.client
@@ -176,7 +237,7 @@ def deploy(paramfile):
             'workers': 0,
             'tag': 'cnvlab',
             'pub_key': '%s/.ssh/id_rsa.pub' % os.environ['HOME'],
-            'pull_secret': get_path('openshift_pull.json'),
+            'pull_secret': 'openshift_pull.json',
             'upstream': False,
             'force': False}
     data.update(paramdata)
@@ -192,8 +253,8 @@ def deploy(paramfile):
     masters = data.get('masters')
     workers = data.get('workers')
     tag = data.get('tag')
-    pub_key = get_path(data.get('pub_key'))
-    pull_secret = get_path(data.get('pull_secret'))
+    pub_key = ssh_path(data.get('pub_key'))
+    pull_secret = pwd_path(data.get('pull_secret'))
     upstream = data.get('upstream')
     force = data.get('force')
     if platform == 'openstack' and (api_ip is None or public_api_ip is None):
@@ -205,7 +266,7 @@ def deploy(paramfile):
     if not os.path.exists(pub_key):
         pprint("Missing public key file %s" % pub_key, color='red')
         sys.exit(1)
-    clusterdir = get_path("clusters/%s" % cluster)
+    clusterdir = pwd_path("clusters/%s" % cluster)
     if not force and os.path.exists(clusterdir):
         pprint("Please Remove existing %s first..." % clusterdir, color='red')
         sys.exit(1)
@@ -215,19 +276,20 @@ def deploy(paramfile):
     if OC is None:
         if not os.path.exists('oc'):
             pprint("Downloading oc in current directory", color='blue')
-            fetch('https://mirror.openshift.com/pub/openshift-v4/clients/oc/latest/%s/oc.tar.gz' % SYSTEM,
-                  get_path('.'))
-        OC = "oc"
+            occmd = "curl -s https://mirror.openshift.com/pub/openshift-v4/clients/oc/latest/%s/oc.tar.gz" % SYSTEM
+            occmd += "| tar zxvf - > %s/oc" % pwd_path('.')
+            call(occmd, shell=True)
+        OC = pwd_path("oc")
     INSTALLER = find_executable('openshift-install')
     if find_executable('openshift-install') is None:
-        if not os.path.exists('openshift-install'):
+        if not os.path.exists(pwd_path('openshift-install')):
             if 'registry.svc.ci.openshift.org' in pull_secret:
-                os.system('get_nightly_installer.sh')
+                get_installer(nightly=True)
             elif upstream:
-                os.system('get_upstream_installer.sh')
+                get_upstream_installer()
             else:
-                os.system('get_upstream_stable.sh')
-        INSTALLER = "openshift-install"
+                get_installer()
+        INSTALLER = pwd_path("openshift-install")
     INSTALLER_VERSION = os.popen('%s version' % INSTALLER).readlines()[0].split(" ")[1].strip()
     if upstream:
         COS_VERSION = "latest"
@@ -252,21 +314,21 @@ def deploy(paramfile):
             pprint("Missing %s. Indicate correct image in your parameters file..." % image, color='red')
             os._exit(1)
     paramdata['image'] = image
-    if not os.path.exists(get_path(clusterdir)):
-        os.mkdir(get_path(clusterdir))
+    if not os.path.exists(clusterdir):
+        os.mkdir(clusterdir)
     data['pub_key'] = open(pub_key).read().strip()
     data['pull_secret'] = re.sub(r"\s", "", open(pull_secret).read())
-    installconfig = config.process_inputfile(cluster, get_path("install-config.yaml"), overrides=data)
+    installconfig = config.process_inputfile(cluster, code_path("install-config.yaml"), overrides=data)
     with open("%s/install-config.yaml" % clusterdir, 'w') as f:
         f.write(installconfig)
     call('%s --dir=%s create manifests' % (INSTALLER, clusterdir), shell=True)
-    for f in [f for f in glob(get_path("customisation/*.yaml"))]:
+    for f in [f for f in glob(code_path("customisation/*.yaml"))]:
         if '99-ingress-controller.yaml' in f:
             replicas = masters if workers == 0 else workers
             role = 'master' if workers == 0 else 'worker'
-            installconfig = config.process_inputfile(cluster, get_path(f),
+            installconfig = config.process_inputfile(cluster, code_path(f),
                                                      overrides={'replicas': replicas, 'role': role})
-            with open("%s/openshift/99-ingress-controller.yaml" % get_path(clusterdir), 'w') as f:
+            with open("%s/openshift/99-ingress-controller.yaml" % code_path(clusterdir), 'w') as f:
                 f.write(installconfig)
         else:
             copy2(f, "%s/openshift" % clusterdir)
@@ -276,7 +338,7 @@ def deploy(paramfile):
         pprint("Deploying helper dhcp node" % image, color='green')
         staticdata.update({'network': network, 'dhcp_image': helper_image, 'prefix': cluster,
                           domain: '%s.%s' % (cluster, domain)})
-        config.plan(cluster, inputfile=get_path('dhcp.yml'), overrides=staticdata)
+        config.plan(cluster, inputfile=code_path('dhcp.yml'), overrides=staticdata)
     if platform in virtplatforms:
         if api_ip is None:
             pprint("You need to define api_ip in your parameters file", color='red')
@@ -411,7 +473,7 @@ def deploy(paramfile):
         sedcmd += ' > %s/bootstrap.ign' % clusterdir
         call(sedcmd, shell=True)
     if platform in virtplatforms:
-        config.plan(cluster, inputfile=get_path('ocp.yml'), overrides=paramdata)
+        config.plan(cluster, inputfile=code_path('ocp.yml'), overrides=paramdata)
         call('%s --dir=%s wait-for bootstrap-complete || exit 1' % (INSTALLER, clusterdir), shell=True)
         todelete = ["%s-bootstrap" % cluster]
         if platform in ['kubevirt', 'openstack', 'vsphere']:
@@ -419,7 +481,7 @@ def deploy(paramfile):
         for vm in todelete:
             k.delete(vm)
     else:
-        config.plan(cluster, inputfile=get_path('ocp_cloud.yml'), overrides=paramdata)
+        config.plan(cluster, inputfile=code_path('ocp_cloud.yml'), overrides=paramdata)
         call('%s --dir=%s wait-for bootstrap-complete || exit 1' % (INSTALLER, clusterdir), shell=True)
         todelete = ["%s-bootstrap" % cluster, "%s-bootstrap-helper" % cluster]
         for vm in todelete:
@@ -437,26 +499,29 @@ def deploy(paramfile):
     pprint("Launching install-complete step. Note it will be retried one extra time in case of timeouts", color='blue')
     call(installcommand, shell=True)
     pprint("Deploying certs autoapprover cronjob", color='blue')
-    call("%s create -f %s" % (OC, get_path('autoapprovercron.yml')), shell=True)
+    call("%s create -f %s" % (OC, code_path('autoapprovercron.yml')), shell=True)
 
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Openshift deployer leveraring kcli library')
-    parser.add_argument('-a', '--action', help='Action to run', choices=('clean', 'deploy', 'scale', 'template'),
-                        default='deploy')
-    parser.add_argument('-f', '--paramfile', help='Parameter file')
-    parser.add_argument('-w', '--workers', help='Number of workers to deploy when scaling', type=int)
+    parser.add_argument('-c', '--clean', help='Clean', action='store_true')
+    parser.add_argument('-s', '--scale', help='Scale', action='store_true')
+    parser.add_argument('-t', '--template', help='Template', action='store_true')
+    parser.add_argument('-w', '--workers', help='Total number of workers (used when scaling)', type=int)
+    parser.add_argument('paramfile', help='Parameter file')
     args = parser.parse_args()
-    action = args.action
+    action_scale = args.scale
+    action_clean = args.clean
+    action_template = args.template
     paramfile = args.paramfile
     if paramfile is not None:
-        paramfile = get_path(paramfile)
+        paramfile = pwd_path(paramfile)
     workers = args.workers
-    if action == 'clean':
+    if action_clean:
         clean(paramfile)
-    elif action == 'scale':
+    elif action_scale:
         scale(paramfile, workers)
-    elif action == 'template':
-        template()
+    elif action_template:
+        template(paramfile)
     else:
         deploy(paramfile)
